Refactor creation of redis pool for tests
